+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -138,6 +138,7 @@ pub(crate) fn get_spendable_sapling_note<P: consensus::Parameters>(
     txid: &TxId,
     index: u32,
     target_height: TargetHeight,
+    include_locked: bool,
 ) -> Result<Option<ReceivedNote<ReceivedNoteId, sapling::Note>>, SqliteClientError> {
     super::common::get_spendable_note(
         conn,
@@ -147,6 +148,7 @@ pub(crate) fn get_spendable_sapling_note<P: consensus::Parameters>(
         ShieldedProtocol::Sapling,
         target_height,
         to_received_note,
+        include_locked,
     )
 }
 
@@ -155,6 +157,7 @@ pub(crate) fn get_spendable_sapling_note<P: consensus::Parameters>(
 /// If the tip shard has unscanned ranges below the anchor height and greater than or equal to
 /// the wallet birthday, none of our notes can be spent because we cannot construct witnesses at
 /// the provided anchor height.
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn select_spendable_sapling_notes<P: consensus::Parameters>(
     conn: &Connection,
     params: &P,
@@ -163,6 +166,7 @@ pub(crate) fn select_spendable_sapling_notes<P: consensus::Parameters>(
     target_height: TargetHeight,
     confirmations_policy: ConfirmationsPolicy,
     exclude: &[ReceivedNoteId],
+    include_locked: bool,
 ) -> Result<Vec<ReceivedNote<ReceivedNoteId, sapling::Note>>, SqliteClientError> {
     super::common::select_spendable_notes(
         conn,
@@ -174,6 +178,7 @@ pub(crate) fn select_spendable_sapling_notes<P: consensus::Parameters>(
         exclude,
         ShieldedProtocol::Sapling,
         to_received_note,
+        include_locked,
     )
 }
 
@@ -522,6 +527,16 @@ pub(crate) mod tests {
     #[test]
     fn spend_fails_on_locked_notes() {
         testing::pool::spend_fails_on_locked_notes::<SaplingPoolTester>()
+    }
+
+    #[test]
+    fn explicit_note_locking() {
+        testing::pool::explicit_note_locking::<SaplingPoolTester>()
+    }
+
+    #[test]
+    fn proposal_level_note_locking() {
+        testing::pool::proposal_level_note_locking::<SaplingPoolTester>()
     }
 
     #[test]