+++ zcash_client_sqlite/src/wallet/transparent.rs
@@ -64,7 +64,10 @@ use crate::{
     AccountRef, AccountUuid, AddressRef, TxRef, UtxoId,
     error::SqliteClientError,
     util::Clock,
-    wallet::{common::tx_unexpired_condition, get_account, mempool_height},
+    wallet::{
+        common::{output_unlocked_condition, tx_unexpired_condition},
+        get_account, mempool_height,
+    },
 };
 
 pub(crate) mod ephemeral;
@@ -1105,6 +1108,7 @@ pub(crate) fn get_spendable_transparent_outputs<P: consensus::Parameters>(
     address: &TransparentAddress,
     target_height: TargetHeight,
     confirmations_policy: ConfirmationsPolicy,
+    include_locked: bool,
 ) -> Result<Vec<WalletUtxo>, SqliteClientError> {
     let mut stmt_utxos = conn.prepare(&format!(
         "SELECT t.txid, u.output_index, u.script,
@@ -1119,11 +1123,13 @@ pub(crate) fn get_spendable_transparent_outputs<P: consensus::Parameters>(
          AND ({}) -- the transaction is mined or unexpired with minconf 0
          AND u.id NOT IN ({}) -- and the output is unspent
          AND ({}) -- exclude likely-spent wallet-internal ephemeral outputs
-         AND ({}) -- exclude immature coinbase outputs",
+         AND ({}) -- exclude immature coinbase outputs
+         AND ({}) -- the output is not locked",
         tx_unexpired_condition_minconf_0("t"),
         spent_utxos_clause(),
         excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts"),
         excluding_immature_coinbase_outputs("t"),
+        output_unlocked_condition("u"),
     ))?;
 
     let addr_str = address.encode(params);
@@ -1141,6 +1147,7 @@ pub(crate) fn get_spendable_transparent_outputs<P: consensus::Parameters>(
         ":target_height": u32::from(target_height),
         ":min_confirmations": min_confirmations,
         ":min_value": u64::from(zip317::MARGINAL_FEE),
+        ":include_locked": include_locked,
     ])?;
 
     let mut utxos = Vec::<WalletUtxo>::new();
@@ -1177,19 +1184,19 @@ pub(crate) fn get_transparent_balances<P: consensus::Parameters>(
     let mut result = HashMap::new();
 
     let mut stmt_address_balances = conn.prepare(&format!(
-        "SELECT u.address, u.value_zat, addresses.key_scope
+        "SELECT u.address, u.value_zat, u.lock_expiry_height, addresses.key_scope
          FROM transparent_received_outputs u
          JOIN accounts ON accounts.id = u.account_id
          JOIN transactions t ON t.id_tx = u.transaction_id
          JOIN addresses ON addresses.id = u.address_id
          WHERE accounts.uuid = :account_uuid
          AND u.value_zat > 0
-         AND ({}) -- the transaction is mined or unexpired with minconf 0
+         AND ({}) -- the output is mined with sufficient confirmations, or is unexpired and minconf is 0
          AND u.id NOT IN ({}) -- and the output is unspent
          AND ({}) -- exclude likely-spent wallet-internal ephemeral outputs",
         tx_unexpired_condition_minconf_0("t"),
         spent_utxos_clause(),
-        excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts")
+        excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts"),
     ))?;
 
     let mut rows = stmt_address_balances.query(named_params![
@@ -1202,6 +1209,7 @@ pub(crate) fn get_transparent_balances<P: consensus::Parameters>(
         let taddr_str: String = row.get("address")?;
         let taddr = TransparentAddress::decode(params, &taddr_str)?;
         let value = Zatoshis::from_nonnegative_i64(row.get("value_zat")?)?;
+        let lock_expiry_height: Option<u32> = row.get("lock_expiry_height")?;
         let key_scope_code: i64 = row.get("key_scope")?;
         let key_scope = KeyScope::decode(key_scope_code)?
             .as_transparent()
@@ -1215,14 +1223,24 @@ pub(crate) fn get_transparent_balances<P: consensus::Parameters>(
         let entry = result.entry(taddr).or_insert((key_scope, Balance::ZERO));
         if value <= zip317::MARGINAL_FEE {
             entry.1.add_uneconomic_value(value)?;
+        } else if lock_expiry_height
+            .iter()
+            .any(|h| *h >= u32::from(target_height))
+        {
+            entry.1.add_locked_value(value)?;
         } else {
             entry.1.add_spendable_value(value)?;
         }
     }
 
+    // Compute pending spendable balance.
+    //
     // Pending spendable balance for transparent UTXOs is only relevant for min_confirmations > 0;
     // with min_confirmations == 0, zero-conf spends are allowed and therefore the value will
-    // appear in the spendable balance and we don't want to double-count it.
+    // appear in the spendable balance and we don't want to also count it as pending.
+    //
+    // For pending value, we can ignore locking considerations until the balance is no longer
+    // pending, so we don't check locking here.
     if min_confirmations > 0 {
         let mut stmt_address_balances = conn.prepare(&format!(
             "SELECT u.address, u.value_zat, addresses.key_scope
@@ -1299,7 +1317,7 @@ pub(crate) fn add_transparent_account_balances(
     };
 
     let mut stmt_account_spendable_balances = conn.prepare(&format!(
-        "SELECT accounts.uuid, SUM(u.value_zat)
+        "SELECT accounts.uuid, u.lock_expiry_height, SUM(u.value_zat)
          FROM transparent_received_outputs u
          JOIN accounts ON accounts.id = u.account_id
          JOIN transactions t ON t.id_tx = u.transaction_id
@@ -1307,10 +1325,10 @@ pub(crate) fn add_transparent_account_balances(
          WHERE ({}) -- the transaction is mined or unexpired with minconf 0
          AND u.id NOT IN ({}) -- and the received txo is unspent
          AND ({}) -- exclude likely-spent wallet-internal ephemeral outputs
-         GROUP BY accounts.uuid",
+         GROUP BY accounts.uuid, lock_expiry_height",
         tx_unexpired_condition_minconf_0("t"),
         spent_utxos_clause(),
-        excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts")
+        excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts"),
     ))?;
 
     let mut rows = stmt_account_spendable_balances.query(named_params![
@@ -1320,7 +1338,8 @@ pub(crate) fn add_transparent_account_balances(
 
     while let Some(row) = rows.next()? {
         let account = AccountUuid(row.get(0)?);
-        let raw_value = row.get(1)?;
+        let lock_expiry_height: Option<u32> = row.get(1)?;
+        let raw_value = row.get(2)?;
         let value = Zatoshis::from_nonnegative_i64(raw_value).map_err(|_| {
             SqliteClientError::CorruptedData(format!("Negative UTXO value {raw_value:?}"))
         })?;
@@ -1331,6 +1350,11 @@ pub(crate) fn add_transparent_account_balances(
             .with_unshielded_balance_mut(|bal| {
                 if value <= zip317::MARGINAL_FEE {
                     bal.add_uneconomic_value(value)
+                } else if lock_expiry_height
+                    .iter()
+                    .any(|h| *h >= u32::from(target_height))
+                {
+                    bal.add_locked_value(value)
                 } else {
                     bal.add_spendable_value(value)
                 }
@@ -1343,7 +1367,7 @@ pub(crate) fn add_transparent_account_balances(
     // TODO (#1592): Ability to distinguish between Transparent pending change and pending non-change
     if min_confirmations > 0 {
         let mut stmt_account_unconfirmed_balances = conn.prepare(&format!(
-            "SELECT accounts.uuid, SUM(u.value_zat)
+            "SELECT accounts.uuid, u.lock_expiry_height, SUM(u.value_zat)
              FROM transparent_received_outputs u
              JOIN accounts ON accounts.id = u.account_id
              JOIN transactions t ON t.id_tx = u.transaction_id
@@ -1362,9 +1386,9 @@ pub(crate) fn add_transparent_account_balances(
              )
              AND u.id NOT IN ({}) -- and the received txo is unspent
              AND ({}) -- exclude likely-spent wallet-internal ephemeral outputs
-             GROUP BY accounts.uuid",
+             GROUP BY accounts.uuid, lock_expiry_height",
             spent_utxos_clause(),
-            excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts")
+            excluding_wallet_internal_ephemeral_outputs("u", "addresses", "t", "accounts"),
         ))?;
 
         let mut rows = stmt_account_unconfirmed_balances.query(named_params![
@@ -1374,7 +1398,8 @@ pub(crate) fn add_transparent_account_balances(
 
         while let Some(row) = rows.next()? {
             let account = AccountUuid(row.get(0)?);
-            let raw_value = row.get(1)?;
+            let lock_expiry_height: Option<u32> = row.get(1)?;
+            let raw_value = row.get(2)?;
             let value = Zatoshis::from_nonnegative_i64(raw_value).map_err(|_| {
                 SqliteClientError::CorruptedData(format!("Negative UTXO value {raw_value:?}"))
             })?;
@@ -1385,6 +1410,11 @@ pub(crate) fn add_transparent_account_balances(
                 .with_unshielded_balance_mut(|bal| {
                     if value <= zip317::MARGINAL_FEE {
                         bal.add_uneconomic_value(value)
+                    } else if lock_expiry_height
+                        .iter()
+                        .any(|h| *h >= u32::from(target_height))
+                    {
+                        bal.add_locked_value(value)
                     } else {
                         bal.add_pending_spendable_value(value)
                     }